updated position farm basic modal callback
